+++ otterdog/models/__init__.py
@@ -15,17 +15,19 @@ from collections.abc import Callable, Iterator, Sequence
 from enum import Enum
 from typing import Any, Protocol, TypeVar, cast
 
-from jsonbender import bend  # type: ignore
+from jsonbender import OptionalS, S, bend  # type: ignore
 
 from otterdog.config import SecretResolver
 from otterdog.jsonnet import JsonnetConfig
 from otterdog.providers.github import GitHubProvider
 from otterdog.utils import (
+    UNSET,
     Change,
     IndentingPrinter,
     T,
     associate_by_key,
     is_different_ignoring_order,
+    is_set_and_valid,
     is_unset,
     multi_associate_by_key,
     patch_to_other,
@@ -34,6 +36,7 @@ from otterdog.utils import (
 )
 
 MT = TypeVar("MT", bound="ModelObject")
+EMT = TypeVar("EMT", bound="EmbeddedModelObject")
 
 
 class FailureType(Enum):
@@ -165,6 +168,107 @@ class LivePatchHandler(Protocol):
 
 
 @dataclasses.dataclass
+class EmbeddedModelObject(ABC):
+    """
+    The abstract base class for embedded model objects.
+    """
+
+    @abstractmethod
+    def validate(self, context: ValidationContext, parent_object: Any) -> None: ...
+
+    def get_patch_to(self, other: EmbeddedModelObject) -> dict[str, Any]:
+        if not isinstance(other, self.__class__):
+            raise ValueError(f"'types do not match: {type(self)}' != '{type(other)}'")
+
+        patch_result = {}
+        for key in self.keys(exclude_unset_keys=True):
+            value = self.__getattribute__(key)
+            other_value = other.__getattribute__(key)
+
+            if is_unset(other_value):
+                continue
+
+            patch_needed, diff = patch_to_other(value, other_value)
+            if patch_needed is True:
+                patch_result[key] = diff
+
+        return patch_result
+
+    @abstractmethod
+    def get_jsonnet_template_function(self, jsonnet_config: JsonnetConfig, extend: bool) -> str | None: ...
+
+    def to_jsonnet(
+        self,
+        printer: IndentingPrinter,
+        jsonnet_config: JsonnetConfig,
+        context: PatchContext,
+        default_object: EmbeddedModelObject,
+    ) -> None:
+        patch = self.get_patch_to(default_object)
+
+        template_function = self.get_jsonnet_template_function(jsonnet_config, False)
+        assert template_function is not None
+
+        printer.print(f"{template_function}()")
+        write_patch_object_as_json(patch, printer)
+
+    @classmethod
+    def all_fields(cls) -> list[dataclasses.Field]:
+        return [field for field in dataclasses.fields(cls)]
+
+    def keys(self, exclude_unset_keys: bool = True) -> list[str]:
+        result = list()
+
+        for field in self.all_fields():
+            if exclude_unset_keys:
+                value = self.__getattribute__(field.name)
+                if not is_unset(value):
+                    result.append(field.name)
+            else:
+                result.append(field.name)
+
+        return result
+
+    def to_model_dict(self) -> dict[str, Any]:
+        result = {}
+
+        for key in self.keys(exclude_unset_keys=True):
+            value = self.__getattribute__(key)
+            result[key] = value
+
+        return result
+
+    @classmethod
+    def from_model_data(cls: type[EMT], data: dict[str, Any]) -> EMT:
+        mapping = cls.get_mapping_from_model()
+        return cls(**bend(mapping, data))  # type: ignore
+
+    @classmethod
+    def get_mapping_from_model(cls) -> dict[str, Any]:
+        return {k: OptionalS(k, default=UNSET) for k in map(lambda x: x.name, cls.all_fields())}
+
+    @classmethod
+    def from_provider_data(cls: type[EMT], org_id: str, data: dict[str, Any]) -> EMT:
+        mapping = cls.get_mapping_from_provider(org_id, data)
+        return cls(**bend(mapping, data))  # type: ignore
+
+    @classmethod
+    def get_mapping_from_provider(cls, org_id: str, data: dict[str, Any]) -> dict[str, Any]:
+        return {k: OptionalS(k, default=UNSET) for k in map(lambda x: x.name, cls.all_fields())}
+
+    @classmethod
+    async def dict_to_provider_data(cls, org_id: str, data: dict[str, Any], provider: GitHubProvider) -> dict[str, Any]:
+        mapping = await cls.get_mapping_to_provider(org_id, data, provider)
+        return bend(mapping, data)
+
+    @classmethod
+    async def get_mapping_to_provider(
+        cls, org_id: str, data: dict[str, Any], provider: GitHubProvider
+    ) -> dict[str, Any]:
+        return {field.name: S(field.name) for field in cls.all_fields() if not is_unset(data.get(field.name, UNSET))}
+
+
+@dataclasses.dataclass
 class ModelObject(ABC):
     """
     The abstract base class for any model object.
@@ -236,6 +340,12 @@ class ModelObject(ABC):
 
             if is_unset(from_value):
                 continue
+
+            if isinstance(to_value, EmbeddedModelObject):
+                to_value = to_value.to_model_dict()
+
+            if isinstance(from_value, EmbeddedModelObject):
+                from_value = from_value.to_model_dict()
 
             if is_different_ignoring_order(to_value, from_value):
                 diff_result[key] = Change(from_value, to_value)
@@ -312,6 +422,10 @@ class ModelObject(ABC):
     def is_nested_model_key(cls, key: str) -> bool:
         return cls.is_nested_model(cls._get_field(key))
 
+    @classmethod
+    def is_embedded_model_key(cls, key: str) -> bool:
+        return cls.is_embedded_model(cls._get_field(key))
+
     @staticmethod
     def is_model_only(field: dataclasses.Field) -> bool:
         return field.metadata.get("model_only", False) is True
@@ -319,6 +433,10 @@ class ModelObject(ABC):
     @staticmethod
     def is_nested_model(field: dataclasses.Field) -> bool:
         return field.metadata.get("nested_model", False) is True
+
+    @staticmethod
+    def is_embedded_model(field: dataclasses.Field) -> bool:
+        return field.metadata.get("embedded_model", False) is True
 
     def get_model_objects(self) -> Iterator[tuple[ModelObject, ModelObject]]:
         yield from []
@@ -449,6 +567,8 @@ class ModelObject(ABC):
             value = self.__getattribute__(key)
             if self.is_nested_model_key(key):
                 result[key] = cast(ModelObject, value).to_model_dict(for_diff, include_nested_models)
+            elif self.is_embedded_model_key(key) and is_set_and_valid(value):
+                result[key] = cast(EmbeddedModelObject, value).to_model_dict()
             else:
                 result[key] = value