+++ otterdog/models/ruleset.py
@@ -13,19 +13,127 @@ import dataclasses
 import re
 from typing import Any, ClassVar, TypeVar, cast
 
-from jsonbender import K, OptionalS, S, bend  # type: ignore
+from jsonbender import F, If, K, OptionalS, S, bend  # type: ignore
 
-from otterdog.models import FailureType, ModelObject, ValidationContext
+from otterdog.jsonnet import JsonnetConfig
+from otterdog.models import (
+    EmbeddedModelObject,
+    FailureType,
+    ModelObject,
+    PatchContext,
+    ValidationContext,
+)
 from otterdog.providers.github import GitHubProvider
 from otterdog.utils import (
     UNSET,
+    IndentingPrinter,
     associate_by_key,
+    is_set_and_present,
     is_set_and_valid,
     is_unset,
     print_warn,
+    write_patch_object_as_json,
 )
 
 RS = TypeVar("RS", bound="Ruleset")
+
+
+@dataclasses.dataclass
+class MergeQueueSettings(EmbeddedModelObject):
+    merge_method: str
+    build_concurrency: int
+    min_group_size: int
+    max_group_size: int
+    wait_time_for_minimum_group_size: int
+    status_check_timeout: int
+    requires_all_group_entries_to_pass_required_checks: bool
+
+    def validate(self, context: ValidationContext, parent_object: Any) -> None:
+        for key in self.keys(False):
+            value = self.__getattribute__(key)
+            if is_unset(value):
+                context.add_failure(
+                    FailureType.ERROR,
+                    f"{parent_object.get_model_header(parent_object)} has not set required parameter "
+                    f"'required_merge_queue.{key}'.",
+                )
+
+        if self.merge_method not in {"MERGE", "SQUASH", "REBASE"}:
+            context.add_failure(
+                FailureType.ERROR,
+                f"{parent_object.get_model_header(parent_object)} has 'required_merge_queue.merge_method' of value "
+                f"'{self.merge_method}', only values ('MERGE' | 'SQUASH' | 'REBASE') are allowed.",
+            )
+
+        for key in [
+            "build_concurrency",
+            "min_group_size",
+            "max_group_size",
+            "wait_time_for_minimum_group_size",
+            "status_check_timeout",
+        ]:
+            value = self.__getattribute__(key)
+            if is_set_and_valid(value):
+                if int(value) < 0:
+                    context.add_failure(
+                        FailureType.ERROR,
+                        f"{parent_object.get_model_header(parent_object)} has 'required_merge_queue.{key}' of value "
+                        f"'{value}' while only non-negative integers are allowed.",
+                    )
+
+    def get_jsonnet_template_function(self, jsonnet_config: JsonnetConfig, extend: bool) -> str | None:
+        return f"orgs.{jsonnet_config.create_merge_queue}"
+
+    @classmethod
+    def get_mapping_from_provider(cls, org_id: str, data: dict[str, Any]) -> dict[str, Any]:
+        mapping = super().get_mapping_from_provider(org_id, data)
+
+        mapping.update(
+            {
+                "build_concurrency": S("max_entries_to_build"),
+                "status_check_timeout": S("check_response_timeout_minutes"),
+                "min_group_size": S("min_entries_to_merge"),
+                "max_group_size": S("max_entries_to_merge"),
+                "wait_time_for_minimum_group_size": S("min_entries_to_merge_wait_minutes"),
+                "requires_all_group_entries_to_pass_required_checks": If(
+                    S("grouping_strategy") == K("ALLGREEN"), K(True), K(False)
+                ),
+            }
+        )
+
+        return mapping
+
+    @classmethod
+    async def get_mapping_to_provider(
+        cls, org_id: str, data: dict[str, Any], provider: GitHubProvider
+    ) -> dict[str, Any]:
+        mapping = super().get_mapping_from_provider(org_id, data)
+
+        mapping.update(
+            {
+                "max_entries_to_build": S("build_concurrency"),
+                "check_response_timeout_minutes": S("status_check_timeout"),
+                "min_entries_to_merge": S("min_group_size"),
+                "max_entries_to_merge": S("max_group_size"),
+                "min_entries_to_merge_wait_minutes": S("wait_time_for_minimum_group_size"),
+                "grouping_strategy": If(
+                    S("requires_all_group_entries_to_pass_required_checks") == K(True), K("ALLGREEN"), K("HEADGREEN")
+                ),
+            }
+        )
+
+        for key in [
+            "build_concurrency",
+            "status_check_timeout",
+            "min_group_size",
+            "max_group_size",
+            "wait_time_for_minimum_group_size",
+            "requires_all_group_entries_to_pass_required_checks",
+        ]:
+            if key in mapping:
+                mapping.pop(key)
+
+        return mapping
 
 
 @dataclasses.dataclass
@@ -37,6 +145,7 @@ class Ruleset(ModelObject, abc.ABC):
     id: int = dataclasses.field(metadata={"external_only": True})
     name: str = dataclasses.field(metadata={"key": True})
     node_id: str = dataclasses.field(metadata={"external_only": True})
+    target: str
     enforcement: str
 
     bypass_actors: list[str]
@@ -68,6 +177,8 @@ class Ruleset(ModelObject, abc.ABC):
     requires_deployments: bool
     required_deployment_environments: list[str]
 
+    required_merge_queue: MergeQueueSettings | None = dataclasses.field(metadata={"embedded_model": True})
+
     _roles: ClassVar[dict[str, str]] = {"5": "RepositoryAdmin", "4": "Write", "2": "Maintain", "1": "OrganizationAdmin"}
     _inverted_roles: ClassVar[dict[str, str]] = {v: k for k, v in _roles.items()}
 
@@ -76,8 +187,16 @@ class Ruleset(ModelObject, abc.ABC):
 
         org_settings = cast(GitHubOrganization, context.root_object).settings
 
-        free_plan = org_settings.plan == "free"
+        if is_set_and_valid(self.target):
+            if self.target not in {"branch", "tag", "push"}:
+                context.add_failure(
+                    FailureType.ERROR,
+                    f"{self.get_model_header(parent_object)} has 'target' of value "
+                    f"'{self.target}', "
+                    f"only values ('branch' | 'tag' | 'push') are allowed.",
+                )
 
+        enterprise_plan = org_settings.plan == "enterprise"
         if is_set_and_valid(self.enforcement):
             if self.enforcement not in {"active", "disabled", "evaluate"}:
                 context.add_failure(
@@ -87,11 +206,11 @@ class Ruleset(ModelObject, abc.ABC):
                     f"only values ('active' | 'disabled' | 'evaluate') are allowed.",
                 )
 
-            if free_plan is True and self.enforcement == "evaluate":
+            if enterprise_plan is False and self.enforcement == "evaluate":
                 context.add_failure(
                     FailureType.ERROR,
                     f"{self.get_model_header(parent_object)} has 'enforcement' of value "
-                    f"'{self.enforcement}' which is not available for a 'free' plan.",
+                    f"'{self.enforcement}' which is only available for an 'enterprise' plan.",
                 )
 
         def valid_condition_pattern(pattern: str) -> bool:
@@ -192,6 +311,9 @@ class Ruleset(ModelObject, abc.ABC):
                         f"'{env_name}' which is not defined in the repository itself.",
                     )
 
+        if is_set_and_present(self.required_merge_queue):
+            self.required_merge_queue.validate(context, parent_object)
+
     def include_field_for_diff_computation(self, field: dataclasses.Field) -> bool:
         # disable diff computation for dependent fields of requires_pull_request,
         if self.requires_pull_request is False:
@@ -223,6 +345,17 @@ class Ruleset(ModelObject, abc.ABC):
     @classmethod
     def from_model_data(cls, data: dict[str, Any]):
         mapping = {k: OptionalS(k, default=UNSET) for k in map(lambda x: x.name, cls.all_fields())}
+
+        mapping.update(
+            {
+                "required_merge_queue": If(
+                    OptionalS("required_merge_queue", default=None) == K(None),
+                    K(None),
+                    S("required_merge_queue") >> F(lambda x: MergeQueueSettings.from_model_data(x)),
+                )
+            }
+        )
+
         return cls(**bend(mapping, data))
 
     @classmethod
@@ -337,6 +470,13 @@ class Ruleset(ModelObject, abc.ABC):
         else:
             mapping["requires_deployments"] = K(False)
             mapping["required_deployment_environments"] = K([])
+
+        # required merge queue
+        if any((found := rule) for rule in rules if rule["type"] == "merge_queue"):
+            merge_queue_parameters = found.get("parameters", {})
+            mapping["required_merge_queue"] = K(MergeQueueSettings.from_provider_data(org_id, merge_queue_parameters))
+        else:
+            mapping["required_merge_queue"] = K(None)
 
         return mapping
 
@@ -500,7 +640,7 @@ class Ruleset(ModelObject, abc.ABC):
 
         pop_mapping(["requires_status_checks", "requires_strict_status_checks", "required_status_checks"])
 
-        # # required deployments
+        # required deployments
         if "requires_deployments" in data:
             value = data["requires_deployments"]
             if value is True:
@@ -514,7 +654,53 @@ class Ruleset(ModelObject, abc.ABC):
 
         pop_mapping(["requires_deployments", "required_deployment_environments"])
 
+        # required merge queue
+        if "required_merge_queue" in data:
+            mapping.pop("required_merge_queue")
+            merge_queue_parameters = await MergeQueueSettings.dict_to_provider_data(
+                org_id, data["required_merge_queue"], provider
+            )
+            if merge_queue_parameters and len(merge_queue_parameters) > 0:
+                rule = {
+                    "type": K("merge_queue"),
+                    "parameters": K(merge_queue_parameters),
+                }
+                rules.append(rule)
+
         if len(rules) > 0:
             mapping["rules"] = rules
 
         return mapping
+
+    def to_jsonnet(
+        self,
+        printer: IndentingPrinter,
+        jsonnet_config: JsonnetConfig,
+        context: PatchContext,
+        extend: bool,
+        default_object: ModelObject,
+    ) -> None:
+        patch = self.get_patch_to(default_object)
+
+        template_function = self.get_jsonnet_template_function(jsonnet_config, False)
+        assert template_function is not None
+
+        if "name" in patch:
+            patch.pop("name")
+
+        printer.print(f"{template_function}('{self.name}')")
+
+        if "required_merge_queue" in patch and patch.get("required_merge_queue") is not None:
+            patch.pop("required_merge_queue")
+
+        write_patch_object_as_json(patch, printer, close_object=False)
+
+        if is_set_and_present(self.required_merge_queue):
+            default_merge_queue_config = MergeQueueSettings.from_model_data(jsonnet_config.default_merge_queue_config)
+            if is_set_and_valid(default_merge_queue_config):
+                printer.print("required_merge_queue: ")
+                self.required_merge_queue.to_jsonnet(printer, jsonnet_config, context, default_merge_queue_config)
+
+        # close the object
+        printer.level_down()
+        printer.println("},")